+++ minecraft/kubejs/server_scripts/custom_recipes.js
@@ -16,7 +16,18 @@ ServerEvents.recipes(event => {
         A: '#minecraft:logs'
     });
 
+    // Recipe for 36 torches from 1 coal block and 1 bundle of sticks
+    event.shaped(Item.of('minecraft:torch', 36), [
+        'A',
+        'B'
+    ], {
+        A: '#c:storage_blocks/coal',
+        B: 'kubejs:bundle_of_sticks'
+    });
+
     event.shapeless('minecraft:leather', Array(9).fill('kubejs:jerky'));
+    event.shapeless('kubejs:block_rotten_flesh', Array(9).fill('minecraft:rotten_flesh'));
+    event.shapeless('kubejs:bundle_of_sticks', Array(9).fill('minecraft:stick'));
 
     // Shapeless decompression recipes
     const decompressions = [
@@ -25,7 +36,9 @@ ServerEvents.recipes(event => {
         { block: 'minecraft:clay', item: 'minecraft:clay_ball', amount: 4 },
         { block: 'minecraft:glowstone', item: 'minecraft:glowstone_dust', amount: 4 },
         { block: '#minecraft:wool', item: 'minecraft:string', amount: 4 },
-        { block: 'minecraft:honeycomb_block', item: 'minecraft:honeycomb', amount: 4 }
+        { block: 'minecraft:honeycomb_block', item: 'minecraft:honeycomb', amount: 4 },
+        { block: 'kubejs:block_rotten_flesh', item: 'minecraft:rotten_flesh', amount: 9 },
+        { block: 'kubejs:bundle_of_sticks', item: 'minecraft:stick', amount: 9 }
     ];
 
     decompressions.forEach(entry => {
@@ -34,9 +47,9 @@ ServerEvents.recipes(event => {
 
     // Furnace smelting recipes
     const smeltables = [
-        { raw: 'minecraft:raw_iron_block', result: 'minecraft:iron_block', xp: 5, cookingTime: 2400 },
-        { raw: 'minecraft:raw_copper_block', result: 'minecraft:copper_block', xp: 5, cookingTime: 2400 },
-        { raw: 'minecraft:raw_gold_block', result: 'minecraft:gold_block', xp: 5, cookingTime: 2400 },
+        { raw: 'minecraft:raw_iron_block', result: 'minecraft:iron_block', xp: 6.3, cookingTime: 1800 },
+        { raw: 'minecraft:raw_copper_block', result: 'minecraft:copper_block', xp: 6.3, cookingTime: 1800 },
+        { raw: 'minecraft:raw_gold_block', result: 'minecraft:gold_block', xp: 9, cookingTime: 1800 },
         { raw: 'minecraft:rotten_flesh', result: 'kubejs:jerky', xp: 0.35, cookingTime: 200 }
     ];